split up Apps into Apps and Overlays
